NR-38213 consolidated all hapi instrumentation into one file. most of the code was no longer needed because we dropped support for < 20 in 9.0.0.  I also moved vision to @hapi/vision as it was missed during our migration of hapi packages
