+++ lib/instrumentations.js
@@ -36,7 +36,7 @@ module.exports = function instrumentations() {
     'restify': { type: MODULE_TYPE.WEB_FRAMEWORK },
     'superagent': { module: '@newrelic/superagent' },
     'undici': { type: MODULE_TYPE.TRANSACTION },
-    'vision': { type: MODULE_TYPE.WEB_FRAMEWORK },
+    '@hapi/vision': { type: MODULE_TYPE.WEB_FRAMEWORK },
     'when': { type: null },
     'winston': { type: MODULE_TYPE.GENERIC }
   }
