+++ test/versioned/hapi/hapi-utils.js
