+++ test/versioned/hapi/newrelic.js
