+++ test/versioned/hapi/plugins.tap.js
@@ -5,7 +5,7 @@
 
 'use strict'
 
-const helper = require('../../../lib/agent_helper')
+const helper = require('../../lib/agent_helper')
 const request = require('request')
 const tap = require('tap')
 const utils = require('./hapi-utils')
